refactor: Convert util 추가
